requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.43.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
faiss-cpu==1.8.0
tiktoken==0.7.0
arize-phoenix==3.19.0
httpx==0.27.0

streamlit==1.38.0
langgraph==0.2.35
langchain==0.2.16